WIP in changes to support new server page

+++ n4d-easysites.install/usr/share/n4d/python-plugins/EasySitesManager.py
@@ -5,16 +5,17 @@ import codecs
 import shutil
 import xmlrpclib as n4dclient
 import ssl
-from jinja2 import Environment
-from jinja2.loaders import FileSystemLoader
-from jinja2 import Template
+#from jinja2 import Environment
+#from jinja2.loaders import FileSystemLoader
+#from jinja2 import Template
 
 class EasySitesManager(object):
 
 	def __init__(self):
 
 		self.config_dir=os.path.expanduser("/etc/easysites/")
-		self.tpl_env = Environment(loader=FileSystemLoader('/usr/share/lliurex-easy-sites/templates'))
+		#self.tpl_env = Environment(loader=FileSystemLoader('/usr/share/lliurex-easy-sites/templates'))
+		self.site_template="/usr/share/lliurex-www/templates/link.json"
 		self.net_folder="/net/server-sync/easy-sites"
 		self.var_folder="/var/www/"
 		self.links_path="/var/lib/lliurex-www/links"
@@ -92,7 +93,7 @@ class EasySitesManager(object):
 					self.sites_config[siteId]=content
 					result=self._update_from_site_link(siteId)
 					if result["status"]:
-						self.sites_config[siteId]["visible"]=result["visible"]
+						self.sites_config[siteId]["visibility"]=result["visibility"]
 						if result["url"]!="":
 							self.sites_config[siteId]["url"]=result["url"]
 						if "http://server/srv/" in self.sites_config[siteId]["image"]["img_path"]:
@@ -121,11 +122,13 @@ class EasySitesManager(object):
 			f=open(site_link)
 			try:
 				content=json.load(f)
-				return {"status":True,"visible":content["visible"],"url":content["link"]}
+				f.close()
+				return {"status":True,"visibility":content["visibility"],"url":content["link"]}
 			except:
-				return {"status":False,"visible":False,"url":""}
+				f.close()
+				return {"status":False,"visibility":False,"url":""}
 		else:
-			return {"status":True,"visible":False,"url":""}
+			return {"status":True,"visibility":False,"url":""}
 	
 	#def _get_site_visibility					
 
@@ -184,7 +187,7 @@ class EasySitesManager(object):
 					result_symlink=self._create_symlink_folder(info["id"])
 					if not result_symlink['status']:
 						'''
-						if not info["visible"]:
+						if not info["visibility"]:
 							result_visible=self._hide_show_site(info["id"],False)
 							if not result_visible['status']:
 								error=True
@@ -263,7 +266,7 @@ class EasySitesManager(object):
 							link_changed=True
 			if not error:	
 				if "visible" in actions_todo:
-					result_visible=self._hide_show_site(info["id"],info["visible"])
+					result_visible=self._hide_show_site(info["id"],info["visibility"])
 					if not result_visible['status']:
 						self._undo_edit_changes(origId,info,rename,icon_changed,link_changed)
 						error=True
@@ -281,7 +284,7 @@ class EasySitesManager(object):
 				else:
 					self._undo_edit_changes(origId,info,rename,icon_changed,link_changed)
 					if visible_changed:
-						if info["visible"]:
+						if info["visibility"]:
 							self._hide_show_site(info["id"],False)
 						else:
 							self._hide_show_site(info["id"],True)	
@@ -342,7 +345,7 @@ class EasySitesManager(object):
 
 		result=self._hide_show_site(info["id"],visible)
 		if result['status']:
-			info['visible']=visible
+			info['visibility']=visible
 			result_write=self.write_conf(info)
 			if not result_write['status']:
 				if visible:
@@ -446,17 +449,16 @@ class EasySitesManager(object):
 			-16:Unable to create link template
 		'''
 		try:
+			new_site=True
 			link_template=os.path.join(self.links_path,"easy-"+info["id"])+".json"
+			'''		
 			site_info={}
 			site_info["ID"]=info["id"]
 			site_info["ICON"]="easy-"+info["id"]+".png"
 			site_info["NAME"]=info["name"]
 			site_info["DESCR"]=info["description"]
-			if info["visible"]:
-				visible="true"
-			else:
-				visible="false"	
-			site_info["VISIBLE"]=visible
+						
+			site_info["visibility"]=visible
 			site_info["EDITABLE"]="false"
 				
 			template= self.tpl_env.get_template("custom.json")
@@ -468,20 +470,44 @@ class EasySitesManager(object):
 			file=open(link_template,"w")
 			file.write(string_template)
 			file.close()
-
+			'''
 			if origId!= None:
-				if info["id"]!=origId:
-					old_link_template=os.path.join(self.links_path,"easy-"+origId)+".json"
-					if os.path.exists(old_link_template):
-						os.remove(old_link_template)
+				current_link=os.path.join(self.links_path,"easy-"+origId)+".json"
+				new_site=False
+			else:
+				current_link=self.site_template
+
+			f=open(current_link)
+			content=json.load(f)
+			content["linkId"]=info["id"]
+			content["link"]="http://server/"+"easy-"+info["id"]
+			content["name"]["default"]=info["name"]
+			content["icon"]="easy-"+info["id"]+".png"
+			content["description"]["default"]=info["description"]
+			content["visibility"]=info["visibility"]
+			if new_site:
+				content["editable"]=False
+				content["order"]=666
+
+			f.close()
+			with codecs.open(link_template,'w',encoding="utf-8") as f:
+				json.dump(content,f,ensure_ascii=False)
+				f.close()
+
+			
+			if info["id"]!=origId and origId!=None:
+				old_link_template=os.path.join(self.links_path,"easy-"+origId)+".json"
+				if os.path.exists(old_link_template):
+					os.remove(old_link_template)
+					'''
 					else:
 						old_hide_link_template=os.path.join(self.hide_folder,"easy-"+origId)+".json"
 						if os.path.exists(old_hide_link_template):
 							os.remove(old_hide_link_template)	
-
-
+					'''
+			
 			result={"status":True,"msg":"Link file create successfuly","code":"","data":""}
-
+		
 		except Exception as e:
 
 			result={"status":False,"msg":str(e),"code":16,"data":""}
@@ -529,19 +555,12 @@ class EasySitesManager(object):
 				pass
 			else:
 				new_symlink="easy-"+siteId	
-				print(new_symlink)
 				new_symlink_path=os.path.join(self.var_folder,new_symlink)
-				print(new_symlink_path)
 				source_symlink_path=os.path.join(self.net_folder,new_symlink)
-				print(source_symlink_path)
 				os.symlink(source_symlink_path,new_symlink_path)
-				print("1")
 				if origId!=None:
-					print("2")
 					old_symlink="easy-"+origId
-					print("2")
 					old_symlink_path=os.path.join(self.var_folder,old_symlink)
-					print("3")
 					os.unlink(old_symlink_path)
 
 			result={"status":True,"msg":"link to /net create successfully","code":"","data":""}
@@ -560,7 +579,6 @@ class EasySitesManager(object):
 			-19:Unable to execute action
 		'''
 		show_site=visible
-		print(show_site)
 		link_site="easy-"+siteId+".json"
 		#hide_site_path=os.path.join(self.hide_folder,link_site)
 		link_site_path=os.path.join(self.links_path,link_site)
@@ -582,7 +600,7 @@ class EasySitesManager(object):
 			
 			f=open(link_site_path)
 			content=json.load(f)
-			content["visible"]=visible
+			content["visibility"]=visible
 			f.close()
 
 			with codecs.open(link_site_path,'w',encoding="utf-8") as f:
@@ -678,7 +696,7 @@ class EasySitesManager(object):
 		if info["id"]!=origId:
 			actions.append("rename")
 
-		if info["visible"]!=self.sites_config[origId]["visible"]:
+		if info["visibility"]!=self.sites_config[origId]["visibility"]:
 			actions.append("visible")
 
 		if info["name"]!=self.sites_config[origId]["name"]:
